# Conflicts:
#	miscroservice_users/src/main/java/com/e_commerce/miscroservice/guanzhao_proj/product_order/dao/GZOrderDao.java
#	miscroservice_users/src/main/java/com/e_commerce/miscroservice/guanzhao_proj/product_order/dao/GZSubjectDao.java
#	miscroservice_users/src/main/java/com/e_commerce/miscroservice/guanzhao_proj/product_order/dao/impl/GZOrderDaoImpl.java
#	miscroservice_users/src/main/java/com/e_commerce/miscroservice/guanzhao_proj/product_order/service/GZSubjectService.java
#	miscroservice_users/src/main/java/com/e_commerce/miscroservice/guanzhao_proj/product_order/service/impl/GZSubjectServiceImpl.java

+++ miscroservice_users/src/main/java/com/e_commerce/miscroservice/guanzhao_proj/product_order/po/TGzVoucher.java
@@ -26,7 +26,7 @@ public class TGzVoucher implements Serializable {
     private Double price;
 
     @Column(commit = "满减下限", length = 11)
-    private Integer reductionLimit;
+    private Double reductionLimit;
 
     @Column(commit = "有效时长(时间戳)")
     private Long effectiveTime;
